Adding font-display to @font-face rule
For a specimen site it would make sense to use `font-display:block`
as there is no point in showing the content in a system font. Using
our own font is the whole point of a specimen site, right? ;-)

But to appease the site speed gods, we put in a `swap` (which goes
straight to the fallback font, and swaps in Fraunces when it has
loaded).

In the end it won't matter to the end-user, as we'll use FFO to
determine when the fonts have loaded, and likely hide the site
content until this has happened. This means that the `font-display`
behaviour is irrelevant as we take showing/hiding content in our
own hands.